remove graphqlbackend dependency from recipients
